# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0",
    "title": "PlooralApiRestTest",
    "description": "API REST Test Plooral"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "LocalTest"
    },
    {
      "url": "http://ec2-54-94-2-165.sa-east-1.compute.amazonaws.com",
      "description": "LocalProd"
    }
  ],
...
  "paths": {
...
    "/v1/companies/": {
      "get": {
        "tags": [
          "companies"
        ],
        "description": "Endpoint to list companies.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "84fec672-f25f-4d92-92eb-fcb80a3c4f1f"
                      },
                      "name": {
                        "type": "string",
                        "example": "ACME Enterprises"
                      },
                      "created_at": {
                        "type": "string",
                        "example": "2024-02-16T13:03:03.424Z"
                      },
                      "updated_at": {
                        "type": "string",
                        "example": "2024-02-16T13:03:03.424Z"
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2024-02-16T13:03:03.424Z"
                      },
                      "updatedAt": {
                        "type": "string",
                        "example": "2024-02-16T13:03:03.424Z"
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "84fec672-f25f-4d92-92eb-fcb80a3c4f1f"
                      },
                      "name": {
                        "type": "string",
                        "example": "ACME Enterprises"
                      },
                      "created_at": {
                        "type": "string",
                        "example": "2024-02-16T13:03:03.424Z"
                      },
                      "updated_at": {
                        "type": "string",
                        "example": "2024-02-16T13:03:03.424Z"
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2024-02-16T13:03:03.424Z"
                      },
                      "updatedAt": {
                        "type": "string",
                        "example": "2024-02-16T13:03:03.424Z"
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Companies were not found!"
                    },
                    "status": {
                      "type": "number",
                      "example": 400
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Companies were not found!"
                    },
                    "status": {
                      "type": "number",
                      "example": 400
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_SERVER_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERROR"
                    },
                    "status": {
                      "type": "number",
                      "example": 500
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERROR"
                    },
                    "status": {
                      "type": "number",
                      "example": 500
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "token": []
          }
        ]
      }
    },
    "/v1/companies/{id}": {
      "get": {
        "tags": [
          "companies"
        ],
        "description": "Endpoint to list a specific company.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "5a655ed8-2cee-4e9b-8fe2-b9c8c884f512"
                      },
                      "name": {
                        "type": "string",
                        "example": "ABC Corp"
                      },
                      "created_at": {
                        "type": "string",
                        "example": "2024-02-16T13:03:03.424Z"
                      },
                      "updated_at": {
                        "type": "string",
                        "example": "2024-02-16T13:03:03.424Z"
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2024-02-16T13:03:03.424Z"
                      },
                      "updatedAt": {
                        "type": "string",
                        "example": "2024-02-16T13:03:03.424Z"
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "5a655ed8-2cee-4e9b-8fe2-b9c8c884f512"
                      },
                      "name": {
                        "type": "string",
                        "example": "ABC Corp"
                      },
                      "created_at": {
                        "type": "string",
                        "example": "2024-02-16T13:03:03.424Z"
                      },
                      "updated_at": {
                        "type": "string",
                        "example": "2024-02-16T13:03:03.424Z"
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2024-02-16T13:03:03.424Z"
                      },
                      "updatedAt": {
                        "type": "string",
                        "example": "2024-02-16T13:03:03.424Z"
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Companies were not found!"
                    },
                    "status": {
                      "type": "number",
                      "example": 400
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Companies were not found!"
                    },
                    "status": {
                      "type": "number",
                      "example": 400
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_SERVER_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERROR"
                    },
                    "status": {
                      "type": "number",
                      "example": 500
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERROR"
                    },
                    "status": {
                      "type": "number",
                      "example": 500
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "token": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "token": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Token access"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Token access"
      }
    }
  }
}